requirements.txt
# Core
boto3>=1.34.0
google-cloud-billing>=1.13.0
azure-mgmt-costmanagement>=4.0.0
requests>=2.31.0
httpx>=0.26.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# CLI
typer>=0.9.0
rich>=13.7.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
